Add ECDSA authentication

+++ flash.c
@@ -184,7 +184,7 @@ static inline int flash_erase(int nr)
 {
 	/* FIXME: make sure that no data in the sector is cached */
 
-	if (nr == FLASH_MASS_ERASE) {
+	if ((unsigned int)nr == FLASH_MASS_ERASE) {
 		flash_erase_all();
 		return get_errflags();
 	}
@@ -264,7 +264,7 @@ retry:
 		if (!overwrite) { /* Save the sector in a temporal sector */
 			tmp = get_temporal_sector_addr(ss);
 
-			if (flash_write_core((void *)tmp, (void *)base, ss, true) != ss)
+			if (flash_write_core((void *)tmp, (void *)base, ss, true) != (size_t)ss)
 				goto out;
 
 			restore = (unsigned int *)tmp;
